ISSUE-166: Add more test coverage
