+++ src/Domain/Strava/Activity/Activity.php
@@ -117,8 +117,8 @@ final class Activity
             sportType: SportType::from($rawData['sport_type']),
             name: $rawData['name'],
             description: $rawData['description'],
-            distance: Kilometer::from($rawData['distance'] / 1000),
-            elevation: Meter::from($rawData['total_elevation_gain']),
+            distance: Kilometer::from(round($rawData['distance'] / 1000, 3)),
+            elevation: Meter::from(round($rawData['total_elevation_gain'])),
             startingCoordinate: Coordinate::createFromOptionalLatAndLng(
                 Latitude::fromOptionalString($rawData['start_latlng'][0] ?? null),
                 Longitude::fromOptionalString($rawData['start_latlng'][1] ?? null),
@@ -126,8 +126,8 @@ final class Activity
             calories: (int) ($rawData['calories'] ?? 0),
             averagePower: isset($rawData['average_watts']) ? (int) $rawData['average_watts'] : null,
             maxPower: isset($rawData['max_watts']) ? (int) $rawData['max_watts'] : null,
-            averageSpeed: KmPerHour::from($rawData['average_speed'] * 3.6),
-            maxSpeed: KmPerHour::from($rawData['max_speed'] * 3.6),
+            averageSpeed: KmPerHour::from(round($rawData['average_speed'] * 3.6, 3)),
+            maxSpeed: KmPerHour::from(round($rawData['max_speed'] * 3.6, 3)),
             averageHeartRate: isset($rawData['average_heartrate']) ? (int) round($rawData['average_heartrate']) : null,
             maxHeartRate: isset($rawData['max_heartrate']) ? (int) round($rawData['max_heartrate']) : null,
             averageCadence: isset($rawData['average_cadence']) ? (int) round($rawData['average_cadence']) : null,
